+++ sources/Device/src/Menu/Pages/Include/PageMemory.h
@@ -4,7 +4,7 @@
 
 
 /// ѕреобразует перечисление в количество точек
-#define ENUM_TO_REL_POINTS(enum)    ((uint)(1 << ((enum) + 9)))
+#define ENUM_TO_REL_POINTS(enum)    ((uint)(1 << (static_cast<int>(enum) + 9)))
 #define FPGA_NUM_POINTS             ENUM_TO_REL_POINTS(set.mem.enumPoints)
